change worker to 0 in gpu mode

+++ data/mazes.py
@@ -104,4 +104,4 @@ class DataReader(object):
         frames = tf.transpose(frames, (0, 1, 3, 2))
         frames = tf.transpose(frames, (0, 2, 1, 3))
         frames = tf.transpose(frames, (1, 0, 2, 3))
-        return frames
+        return tf.cast(frames, tf.float32)
